refac: refactory decoder

+++ src/rs_driver/driver/driver_param.h
@@ -214,7 +214,7 @@ typedef struct RSDecoderParam  ///< LiDAR decoder parameter
                                                                      ///< 2: Split frames by fixed number of packets;
                                                                      ///< 3: Split frames by custom number of packets (num_pkts_split)
   float split_angle = 0.0f;        ///< Cut angle(degree) used to split frame, only be used when split_frame_mode=1
-  uint32_t num_blks_split = 1;     ///< Number of packets in one frame, only be used when split_frame_mode=3
+  uint16_t num_blks_split = 1;     ///< Number of packets in one frame, only be used when split_frame_mode=3
   bool wait_for_difop = true;      ///< true: start sending point cloud until receive difop packet
   bool use_lidar_clock = false;    ///< true: use LiDAR clock as timestamp; false: use system clock as timestamp
   bool dense_points = false;  ///< true: discard NAN points; false: reserve NAN points
